reorganize data packages
