+++ data/event/event_test.go
@@ -1,4 +1,4 @@
-package metric
+package event
 
 import (
 	"testing"
