+++ data/inventory/inventory_test.go
@@ -1,4 +1,4 @@
-package metric
+package inventory
 
 import (
 	"testing"
