+++ integration/entity_test.go
@@ -8,7 +8,7 @@ import (
 
 	"encoding/json"
 
-	"github.com/newrelic/infra-integrations-sdk/metric"
+	"github.com/newrelic/infra-integrations-sdk/data/metric"
 	"github.com/newrelic/infra-integrations-sdk/persist"
 	"github.com/stretchr/testify/assert"
 )
